+++ Assets/cosine_table.h
@@ -1,4 +1,4 @@
 #define COSINE_TABLE_LEN 512
 
-extern const int tcos[COSINE_TABLE_LEN];
-extern const int tsin[COSINE_TABLE_LEN];
+extern const short tcos[COSINE_TABLE_LEN];
+extern const short tsin[COSINE_TABLE_LEN];
